Crate system

Re-exports

pub use custom_type::CopyPath;

pub use custom_type::PathType;

pub use custom_type::ClonePath;

pub use errors::SystemError;

Modules

custom_type

errors

Functions

create_hash

Create a 256-bit hash for the given data.

del_dir

Deletes a directory.

del_file

Deletes a file.

generate_random_string

Generates a random string of the specified length using alphanumeric characters.

is_pathDeprecated

Deprecated function. Use path_present instead.

is_string_in_file

Checking if file contains a specific string.

make_dir

Creates a directory if it does not exist.

make_dir_perm

Creates a directory with the specified permissions.

make_file

Creates a file if it does not exist.

path_present

Checks if a path exists.

remake_dir

Recreates a directory by first deleting it if it exists and then creating it.

truncate

Trims a string to a maximum number of characters.

untar

Extracts the contents of a tar.gz file to a specified output folder.