Module system::scheme

Reexports

use core::prelude::v1::*;

use core::ops::{Deref, DerefMut};

use core::{mem, ptr, slice, str};

use super::error::*;

use super::syscall::*;

Structs

Packet

Traits

Scheme

Functions

c_string_to_str
Helper function for handling C strings, please do not copy it or make it pub or change it